+++ lib/prism_gradients_toolkit.dart
@@ -1,4 +1,4 @@
-library prism_gradient_toolkit;
+library prism_gradients_toolkit;
 
 // Export your widgets here
 export 'src/gradient avatar/gradient_avatar.dart';